Backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==22.0.0
httpx==0.27.2
PyJWT==2.9.0
pydantic==2.7.4
python-multipart==0.0.12
python-dotenv==1.0.1 